Moved all interfaces and enums to /shared in seperate files
