+++ shared/enums.ts
@@ -1,0 +1,10 @@
+export enum OSS_TRANSLATIONS {
+    openShop = 'Open Shop',
+    openSellingShop = 'Open Shop',
+    notEnoughCash = 'Not enough cash!',
+}
+
+export enum ShopType {
+    BUY = 'buy', //Players can buy stuff
+    SELL = 'sell', //Players can sell stuff
+}
