--- shared/interfaces/IShopItem.ts
@@ -1,6 +1,0 @@
-export interface iShopItem {
-    dbName: string;
-    price: number;
-    image: string,
-    name: string,
-}
